[Refactor] #208 Coordinator 개선
